+++ crates/fluvio-smartengine/src/engine/wasmtime_engine/mod.rs
@@ -4,6 +4,6 @@ pub(crate) mod init;
 pub(crate) mod state;
 pub(crate) mod engine;
 pub(crate) mod instance;
-pub use engine::{SmartEngineImp, initialize_imp, SmartModuleChainInstanceImp};
+pub use engine::{SmartEngine, SmartModuleChainBuilder, SmartModuleChainInstance};
 
 use super::*;
