+++ crates/fluvio-smartengine/src/engine/wasmtime_engine/transforms/aggregate.rs
@@ -133,8 +133,7 @@ mod test {
 
         let mut chain = chain_builder
             .initialize(&engine)
-            .expect("failed to build chain")
-            .inner;
+            .expect("failed to build chain");
 
         assert_eq!(
             chain.instances().first().expect("first").transform().name(),
