+++ crates/fluvio-smartengine/src/engine/wasmtime_engine/transforms/filter.rs
@@ -31,8 +31,7 @@ mod test {
 
         let mut chain = chain_builder
             .initialize(&engine)
-            .expect("failed to build chain")
-            .inner;
+            .expect("failed to build chain");
 
         assert_eq!(
             chain.instances().first().expect("first").transform().name(),
@@ -90,8 +89,7 @@ mod test {
 
         let mut chain = chain_builder
             .initialize(&engine)
-            .expect("failed to build chain")
-            .inner;
+            .expect("failed to build chain");
 
         let instance = chain.instances().first().expect("first");
